+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     * Fixed anchor point detection in imported SVG
     * Added support for split positions in imported lotties
     * Fixed SVG export of precomps with stretched/offset time
+    * Added support for loading image assets from the web
 * Scripting
     * `window.choose_option()` method that shows a dilaog to display a selection
     * Exception from I/O plugins now properly show in the script console